+++ src/game-state/src/common-item.h
@@ -2,6 +2,7 @@
 #define _COMMON_ITEM_H
 
 #include "item.h"
+#include "game_action.h"
 
 /* item_init() initializes an item struct with given values
     arguments are taken from WDL
@@ -13,8 +14,7 @@
   Returns:
     FAILURE for failure, SUCCESS for success
 */
-int item_init(item_t *new_item, char *item_id, char *short_desc,
-    char *long_desc);
+int item_init(item_t *new_item, char *item_id, char *short_desc, char *long_desc);
 
 
 /* this has to be in the interface as room and player modules use this */
@@ -63,7 +63,7 @@ int delete_all_attributes(attribute_hash_t attributes);
 *  Returns:
 *    A pointer to a new action struct.
 */
-game_action_t *game_action_new(char *act_name, action_type_t *act_type);
+game_action_t *game_action_new(char *act_name, char* success_str, char* fail_str);
 
 /* action_init() initializes an action struct with given values
    arguments are taken from action management
@@ -74,6 +74,5 @@ game_action_t *game_action_new(char *act_name, action_type_t *act_type);
  Returns:
     FAILURE for failure, SUCCESS for success
 */
-int game_action_init(game_action_t *new_action, char *act_name,
-    action_type_t *act_type);
+int game_action_init(game_action_t *new_action, char *act_name, char* success_str, char* fail_str);
 #endif
